+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 - `firebase emulator:start` use a default project `demo-no-project` if no project can be found. (#9072)
 - `firebase init dataconnect` also supports bootstrapping flutter template. (#9084)
+- `firebase init dataconnect` confirms Cloud SQL provisioning. (#9095)
 - MCP `firebase_init` tool can download an existing FDC service returned from `dataconnect_list_services` (#9091)